Test slug generation
